+++ docs/content/examples/grid/MultipleSelection.js
@@ -1,10 +1,12 @@
-import { HtmlElement, Checkbox, Grid, Tab } from 'cx/widgets';
+import { HtmlElement, Checkbox, Grid, Tab, TextField } from 'cx/widgets';
 import { Content, Controller, PropertySelection } from 'cx/ui';
 import {Md} from '../../../components/Md';
 import {CodeSplit} from '../../../components/CodeSplit';
 import {CodeSnippet} from '../../../components/CodeSnippet';
 
 import {casual} from '../data/casual';
+import { getSearchQueryPredicate } from 'cx/util';
+import { updateArray } from 'cx/data';
 
 class PageController extends Controller {
     init() {
@@ -20,18 +22,29 @@ class PageController extends Controller {
 
         this.addTrigger('select-all-click', ['$page.selectAll'], (value) => {
             if (value != null)
-                this.store.set('$page.records', this.store.get('$page.records')
-                    .map(r => Object.assign({}, r, {selected: value})));
+                this.store.update('$page.records', updateArray, r => ({ ...r, selected: value }), r => this.visibleIdsMap[r.id]);
         });
+    }
 
-        this.addTrigger('item-click', ['$page.records'], (records) => {
-            if (records.every(a => a.selected))
-                this.store.set('$page.selectAll', true);
-            else if (records.every(a => !a.selected))
-                this.store.set('$page.selectAll', false);
-            else
-                this.store.set('$page.selectAll', null);
-        }, true);
+    updateSelection(newRecords, instance) {
+        this.visibleIdsMap = newRecords.reduce((acc, r) => {
+            acc[r.data.id] = true;
+            return acc;
+        }, {});
+
+        let anySelected, anyUnselected;
+        for (let rec of newRecords) {
+            if (rec.data.selected === true)
+                anySelected = true;
+
+            else if (rec.data.selected === false)
+                anyUnselected = true;
+
+            if (anySelected && anyUnselected)
+                break;
+        }
+
+        this.store.set('$page.selectAll', anySelected && anyUnselected ? null : !!anySelected);
     }
 }
 
@@ -46,6 +59,13 @@ export const MultipleSelection = <cx>
             or checkboxes.
 
             To select all rows click the checkbox in the header.
+
+            Example also showcases use of `onTrackMappedRecords` in combination with `onCreateFilter` callback.
+            Using `onTrackMappedRecords` we can access filtered records.
+
+            <div style="margin-bottom: 10px" ws>
+                <TextField value-bind='$page.searchText' icon='search' placeholder='Search...' showClear style='width: 300px'/>
+            </div>
 
             <Grid
                 records-bind='$page.records'
@@ -62,6 +82,13 @@ export const MultipleSelection = <cx>
                 ]}
                 selection={{type: PropertySelection, bind: "$page.selection", multiple: true}}
                 sorters-bind="$page.sorters"
+                filterParams-bind='$page.searchText'
+                onCreateFilter={(searchText) => {
+                    if (!searchText) return () => true;
+                    let predicate = getSearchQueryPredicate(searchText);
+                    return record => predicate(record.fullName) || predicate(record.phone) || predicate(record.city);
+                }}
+                onTrackMappedRecords='updateSelection'
             />
 
             See also:
@@ -84,18 +111,29 @@ export const MultipleSelection = <cx>
 
                             this.addTrigger('select-all-click', ['$page.selectAll'], (value) => {
                                 if (value != null)
-                                    this.store.set('$page.records', this.store.get('$page.records')
-                                        .map(r => Object.assign({}, r, {selected: value})));
+                                    this.store.update('$page.records', updateArray, r => ({ ...r, selected: value }), r => this.visibleIdsMap[r.id]);
                             });
+                        }
 
-                            this.addTrigger('item-click', ['$page.records'], (records) => {
-                                if (records.every(a => a.selected))
-                                    this.store.set('$page.selectAll', true);
-                                else if (records.every(a => !a.selected))
-                                    this.store.set('$page.selectAll', false);
-                                else
-                                    this.store.set('$page.selectAll', null);
-                            }, true);
+                        updateSelection(newRecords, instance) {
+                            this.visibleIdsMap = newRecords.reduce((acc, r) => {
+                                acc[r.data.id] = true;
+                                return acc;
+                            }, {});
+
+                            let anySelected, anyUnselected;
+                            for (let rec of newRecords) {
+                                if (rec.data.selected === true)
+                                    anySelected = true;
+
+                                else if (rec.data.selected === false)
+                                    anyUnselected = true;
+
+                                if (anySelected && anyUnselected)
+                                    break;
+                            }
+
+                            this.store.set('$page.selectAll', anySelected && anyUnselected ? null : !!anySelected);
                         }
                     }`}
                 </CodeSnippet>
@@ -115,6 +153,13 @@ export const MultipleSelection = <cx>
                         ]}
                         selection={{type: PropertySelection, bind: "$page.selection", multiple: true}}
                         sorters-bind="$page.sorters"
+                        filterParams-bind='$page.searchText'
+                        onCreateFilter={(searchText) => {
+                            if (!searchText) return () => true;
+                            let predicate = getSearchQueryPredicate(searchText);
+                            return record => predicate(record.fullName) || predicate(record.phone) || predicate(record.city);
+                        }}
+                        onTrackMappedRecords='updateSelection'
                     />`}
                 </CodeSnippet>
             </Content>