+++ docs/content/widgets/configs/NumberField.js
@@ -109,6 +109,12 @@ export default {
             Largest allowed number value.
         </Md></cx>
     },
+    constrain: {
+        type: 'boolean',
+        description: <cx><Md>
+            When specified, values lower than `minValue` or higher than `maxValue` will be constrained to `minValue` that is, `maxValue`, respectively.
+        </Md></cx>
+    },
     emptyValue: {
         type: 'any',
         description: <cx><Md>