+++ litmus/index.js
@@ -99,12 +99,14 @@ import "./index.scss";
 //import Demo from "./bugs/AsyncValidator";
 
 //import Demo from "./features/charts/pie/labels";
-import Demo from "./bugs/BlurBug";
+//import Demo from "./bugs/BlurBug";
 //import Demo from "./features/immer";
 // import Demo from "./bugs/971";
 // import Demo from "./bugs/LookupFieldListItemFocus";
 //import Demo from "./bugs/SliderValue";
 //import Demo from "./features/context-menu/conext-menu-in-a-dropdown";
+// import Demo from "./features/calendar";
+import Demo from "./features/slider/SliderPreventDefault";
 
 let store = (window.store = new Store());