+++ packages/cx/src/ui/createFunctionalComponent.d.ts
@@ -1,1 +1,1 @@
-export function createFunctionalComponent(factory);
+export function createFunctionalComponent<T>(factory: (props: T) => any): (props: T) => any;