+++ packages/cx/src/util/isNonEmptyArray.d.ts
@@ -1,1 +1,1 @@
-export function isNonEmptyArray(x: any): boolean;
+export function isNonEmptyArray(x: any): x is [any, ...any];